data 분리
