+++ ParserCoroutine.py
@@ -1,53 +1,19 @@
-from enum import Enum
+from data import (
+    ParserMessage,
+    MutableParserState,
+    ImmutablePainState,
+    PainMicrowaveMode,
+    PainLowPulseMode,
+    PainLowPulsePulse,
+)
 from typing import Generator
-from dataclasses import dataclass
-
-class ParserMessage(Enum):
-    FAIL = "[이전 작업 오류] 시작 문자 받는 중..."
-    START = "시작 문자 받는 중..."
-    MICROWAVE_MODE = "초음파 모드 받는 중..."
-    MICROWAVE_VALUE = "초음파 출력값 받는 중..."
-    LOWPULSE_MODE   = "저주파 모드 받는 중..."
-    LOWPULSE_VALUE  = "저주파 출력값 받는 중..."
-    LOWPULSE_PULSE_VALUE = "저주파 펄스 값 받는 중..."
-    LOWPULSE_V_VALUE = "저주파 전류값 받는 중..."
-
-class PainMicrowaveMode(Enum):
-    CONTINUOUS = b'0'[0]
-    PULSE = b'1'[0]
-
-class PainLowPulseMode(Enum):
-    CONTINUOUS = b'0'[0]
-    INTERVAL = b'1'[0]
-
-class PainLowPulsePulse(Enum):
-    PULSE_182Hz_500uS = b'0'[0]
-    PULSE_182Hz_250uS = b'1'[0]
-
-@dataclass(frozen=False)
-class MutableParserState:
-    message: ParserMessage = ParserMessage.START
-    microwave_mode: PainMicrowaveMode = PainMicrowaveMode.CONTINUOUS
-    microwave: int = 0
-    lowpulse_mode: PainLowPulseMode = PainLowPulseMode.CONTINUOUS
-    lowpulse: int = 0
-    lowpulse_pulse: PainLowPulsePulse = PainLowPulsePulse.PULSE_182Hz_500uS
-    lowpulse_v: int = 0
-
-@dataclass(frozen=True)
-class ImmutablePainState:
-    microwave_mode: PainMicrowaveMode = PainMicrowaveMode.CONTINUOUS
-    microwave: int = 0
-    lowpulse_mode: PainLowPulseMode = PainLowPulseMode.CONTINUOUS
-    lowpulse: int = 0
-    lowpulse_pulse: PainLowPulsePulse = PainLowPulsePulse.PULSE_182Hz_500uS
-    lowpulse_v: int = 0
 
 
 def flatten(data: Generator[bytes, None, None]):
     for byte_array in data:
         for byte_item in byte_array:
             yield byte_item
+
 
 def _next_message(message: ParserMessage):
     match message:
@@ -67,6 +33,7 @@ def _next_message(message: ParserMessage):
             return ParserMessage.LOWPULSE_V_VALUE
         case ParserMessage.LOWPULSE_V_VALUE:
             return ParserMessage.START
+
 
 def _update_state(state: MutableParserState, next: int):
     match state.message:
@@ -115,11 +82,12 @@ def _update_state(state: MutableParserState, next: int):
                 state.message = _next_message(state.message)
             else:
                 state.message = ParserMessage.FAIL
-        case _: # FAIL, START
+        case _:  # FAIL, START
             if next == 2:
                 state.message = _next_message(state.message)
             else:
                 state.message = ParserMessage.FAIL
+
 
 def flatmap_parse(data: Generator[int, None, None]):
     state = MutableParserState()
@@ -134,5 +102,5 @@ def flatmap_parse(data: Generator[int, None, None]):
                 state.lowpulse_mode,
                 state.lowpulse,
                 state.lowpulse_pulse,
-                state.lowpulse_v
+                state.lowpulse_v,
             )
